Added Sessions
When a user logins in they are now redirected to the TA page and a session is started. When a session is started the password is removed from the session for security reasons. In the future need to change this so it redirects the user to the appropriate page based on their role.