Role Based Calendar Form
Now when a cell is selected on the calendar we check to see the role of the user. If the user has a role teaching assistant the availability form is opened, whereas module leaders have the request support form opened. Changed also made the the close_form() function to handle this.